Add get methods to the driver-lib
This patch adds get methods to the driver-lib which allows provider
drivers to query for objects by ID.
In support of the get methods, this patch fixes a bug in the data
model to_dict() where it may not properly recurse the objects.
It also improves connecting to the driver agent sockets and adds a
timeout while waiting to receive data from the driver-agent.

Change-Id: Ia69d1f61571a1a65dee585037affb317999d7ade
Story: 2005870
Task: 33682